// src/core/roles/generated.ts
    export enum roles {
      superadmin = 'superadmin', // reserved for system administrator
      site = 'site', // Root for site users
      user = 'user', // Active user
      visitor = 'visitor', // Not confirmed user
      author = 'author', // Author
      admin = 'admin', // Root for site administrators
      supervisor = 'supervisor', // Activity supervisor
      usersadmin = 'usersadmin', // Users administrator
      rolesadmin = 'rolesadmin', // Roles Manager
      default = 'visitor', // default role
    }
    